core/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
frozenlist==1.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
multidict==5.2.0
nltk==3.6.5
oauthlib==3.1.1
pymongo==4.0.1
python-dotenv==0.19.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
textblob==0.17.1
tqdm==4.62.3
tweepy==4.4.0
typing-extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
